api/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.10.0
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
fastapi==0.117.1
flatbuffers==25.2.10
h11==0.16.0
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
numpy==2.2.6
onnxruntime==1.22.1
onnxruntime-gpu==1.23.2
opencv-python==4.12.0.88
packaging==25.0
pefile==2023.2.7
pillow==11.3.0
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.8
pyreadline3==3.5.4
pywin32-ctypes==0.2.3
sniffio==1.3.1
starlette==0.48.0
sympy==1.14.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.36.0
